Reorganize before additional commands are introduced
