+++ includes/wp-cli.php
@@ -35,5 +35,6 @@ function stop_the_insanity() {
 /**
  * Load commands
  */
-require __DIR__ . '/wp-cli/class-data.php';
+require __DIR__ . '/wp-cli/class-cache.php';
+require __DIR__ . '/wp-cli/class-events.php';
 require __DIR__ . '/wp-cli/class-one-time-fixers.php';
